==> babybirth/scripts/02-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves the data from 
# https://dasl.datadescription.com/datafile/babysamp-98/
# Author: ruiyang pang
# Date: 13 Nov 2024
# Contact: ruiyang.pang@mail.utoronto.ca
# License: MIT
# Pre-requisites: [...UPDATE THIS...]
# Any other information needed? [...UPDATE THIS...]


#### Workspace setup ####
library(tidyverse)

#### Download data ####
## directly download the txt file from the website
##  https://dasl.datadescription.com/datafile/babysamp-98/


#### Save data ####
# and save the babysamp-98.txt into the folder data/01-raw_data/
